TCAP ISM: trap the EXIT signal to avoid getting killed too early
When we receive a TC-END with RESULT-L, the process hierarchy might
kill us before we received the RESSULT-L, or aftger we received it,
but before we had a chance to deliver it to the TC User.

Trapping the EXIT flag solves the problem.  It also means that the EXIT
signal will not propagate to other linked processes.  However, there
should be no other processes, as ISM is not starting any.